domain-ssl-tracker/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
python-whois==0.9.4
python-dotenv==1.0.1
apscheduler==3.10.4
jinja2==3.1.4
python-multipart==0.0.9
aiofiles==23.2.1
